requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
motor==3.3.1
pydantic==1.10.12
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
fastapi-mail==1.3.1
structlog==23.1.0
bcrypt==4.0.1
pymongo==4.6.1
